added sequence input function
was having trouble with inputting fraction, so instead of scanning each input as an integer I scanned as string then parsed and used atof

+++ arithmetic.c
@@ -1,7 +1,7 @@
 #include "pattern_header.h"
 
-int is_arithmetic(int arr[], int size) {
-    int difference = arr[1] - arr[0];
+float is_arithmetic(float arr[], int size) {
+    float difference = arr[1] - arr[0];
 
     for (int i = 0; i < size - 1; i++) {
         if (arr[i+1] - arr[i] == difference) {
